Keep print-mode spawn on the Worker Adapter, not as a fourth factory.
