+++ src/adapters/claude.ts
@@ -1,6 +1,5 @@
-import { codingCliWorker } from "./coding-cli.ts";
-import type { WorkerAdapter } from "../worker-adapter.ts";
+import { printModeWorker, type WorkerAdapter } from "../worker-adapter.ts";
 
 export function claude(): WorkerAdapter {
-  return codingCliWorker("claude", "--dangerously-skip-permissions");
+  return printModeWorker("claude", "--dangerously-skip-permissions");
 }
