+++ src/adapters/cursor.ts
@@ -1,6 +1,5 @@
-import { codingCliWorker } from "./coding-cli.ts";
-import type { WorkerAdapter } from "../worker-adapter.ts";
+import { printModeWorker, type WorkerAdapter } from "../worker-adapter.ts";
 
 export function cursor(): WorkerAdapter {
-  return codingCliWorker("agent", "--yolo");
+  return printModeWorker("agent", "--yolo");
 }
