+++ src/worker-adapter.ts
@@ -44,3 +44,20 @@ export function spawnWorkerBinary(
     });
   });
 }
+
+export function printModeWorker(
+  bin: string,
+  unattendedFlag: string,
+): WorkerAdapter {
+  return createWorkerAdapter({
+    bin,
+    spawn(request: SpawnRequest) {
+      const args = ["-p", "--model", request.model];
+      if (request.permissions === "unattended") {
+        args.push(unattendedFlag);
+      }
+      args.push(request.prompt);
+      return spawnWorkerBinary(bin, args, request.cwd);
+    },
+  });
+}
